use stores in plan, move external to lib
