+++ src/lib/fireworks.js
@@ -1,0 +1,36 @@
+// https://www.kirilv.com/canvas-confetti/#fireworks
+import confetti from "canvas-confetti";
+
+export default function fireworks() {
+  var duration = 3 * 1000;
+  var animationEnd = Date.now() + duration;
+  var defaults = {
+    startVelocity: 30,
+    spread: 360,
+    ticks: 60,
+    zIndex: 0,
+    disableForReducedMotion: true,
+  };
+  function randomInRange(min, max) {
+    return Math.random() * (max - min) + min;
+  }
+  var interval = setInterval(function () {
+    var timeLeft = animationEnd - Date.now();
+    if (timeLeft <= 0) {
+      return clearInterval(interval);
+    }
+    var particleCount = 100 * (timeLeft / duration);
+    confetti(
+      Object.assign({}, defaults, {
+        particleCount,
+        origin: { x: randomInRange(0.1, 0.3), y: Math.random() - 0.2 },
+      })
+    );
+    confetti(
+      Object.assign({}, defaults, {
+        particleCount,
+        origin: { x: randomInRange(0.7, 0.9), y: Math.random() - 0.2 },
+      })
+    );
+  }, 250);
+}
